refactor(sponsors): extract sponsor schema into useSponsorSchema composable
The Sponsoring component shipped 25 lines of JSON-LD construction
inline. Move them into a useSponsorSchema composable that takes a ref
to the sponsor list and emits the Organization entities + the sponsor
relationship on OneLiteFeather's identity. The component now wires it
up with a single call, keeping the view file focused on the visible
carousel and CTA.

Output unchanged — the same two Organization entries plus the
sponsor-array patch on the global identity.
