+++ components/features/sponsoring/Sponsoring.vue
@@ -32,30 +32,9 @@ const displaySubtitle = computed(() => props.subtitle ?? t('sponsor.subtitle'))
 
 const enhancedSponsors = computed<Sponsor[]>(() => props.sponsors ?? [])
 
-// Describe each sponsor as a schema.org Organization and attach the list
-// as a `sponsor` array on the OneLiteFeather org (referenced by @id), so
-// Google understands the partnerships beyond the marketing copy.
-const site = useSiteConfig()
-useSchemaOrg(() => {
-  const sponsors = enhancedSponsors.value
-  if (!sponsors.length) return []
-  const organizations = sponsors.map((s) => ({
-    '@type': 'Organization' as const,
-    '@id': sponsorId(site.url, s.name),
-    name: s.name,
-    url: s.url,
-    description: s.description || undefined,
-    logo: s.logo || undefined
-  }))
-  return [
-    ...organizations,
-    {
-      '@type': 'Organization' as const,
-      '@id': organizationId(site.url),
-      sponsor: organizations.map((o) => ({ '@id': o['@id'] }))
-    }
-  ]
-})
+// Schema.org markup for the sponsor list lives in its own composable so
+// this component stays focused on rendering.
+useSponsorSchema(enhancedSponsors)
 
 const resolveIcon = (icon?: Sponsor['icon']) => {
   if (!icon) return null
